requirements.txt
anthropic>=0.40.0
openai>=1.30.0
pymupdf>=1.24.0
pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
streamlit>=1.35.0
